addressing review comment
