+++ catalog-rest-service/src/main/resources/ui/src/components/common/table-data-card/TableDataCard.tsx
@@ -17,13 +17,8 @@
 
 import React, { FunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
-import {
-  getDatasetDetailsPath,
-  getTopicDetailsPath,
-} from '../../../constants/constants';
-import { SearchIndex } from '../../../pages/explore/explore.interface';
 import { stringToHTML } from '../../../utils/StringsUtils';
-import { getUsagePercentile } from '../../../utils/TableUtils';
+import { getEntityLink, getUsagePercentile } from '../../../utils/TableUtils';
 import TableDataCardBody from './TableDataCardBody';
 
 type Props = {
@@ -60,24 +55,11 @@ const TableDataCard: FunctionComponent<Props> = ({
     { key: 'Tier', value: tier },
   ];
 
-  const getEntityLink = (indexType: string) => {
-    switch (indexType) {
-      case SearchIndex.TABLE:
-        return getDatasetDetailsPath(fullyQualifiedName);
-
-      case SearchIndex.TOPIC:
-        return getTopicDetailsPath(fullyQualifiedName);
-
-      default:
-        return getDatasetDetailsPath(fullyQualifiedName);
-    }
-  };
-
   return (
     <div className="tw-bg-white tw-p-3 tw-border tw-border-main tw-rounded-md">
       <div>
         <h6 className="tw-flex tw-items-center tw-m-0 tw-heading">
-          <Link to={getEntityLink(indexType)}>
+          <Link to={getEntityLink(indexType, fullyQualifiedName)}>
             <button className="tw-text-grey-body tw-font-medium">
               {stringToHTML(name)}
             </button>
