+++ catalog-rest-service/src/main/resources/ui/src/enums/search.enum.ts
@@ -20,3 +20,8 @@ export enum FilterType {
   PLATFORM = 'platform',
   CLUSTER = 'cluster',
 }
+
+export enum SearchIndex {
+  TABLE = 'table_search_index',
+  TOPIC = 'topic_search_index',
+}
