+++ catalog-rest-service/src/main/resources/ui/src/pages/explore/explore.interface.ts
@@ -33,8 +33,3 @@ export type Team = {
   description: string;
   href: string;
 };
-
-export enum SearchIndex {
-  TABLE = 'table_search_index',
-  TOPIC = 'topic_search_index',
-}
